Issue #AI4IDS-2365 | feat(ppu): bill audio services (ASR, diarization, lang-detection) on fractional minutes

+++ services/platform-core-service/app/models/pay_per_use/ppu_tier.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import BigInteger, Boolean, Column, DateTime, ForeignKey, String, Text, UniqueConstraint
+from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Numeric, String, Text, UniqueConstraint
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
@@ -50,8 +50,8 @@ class PPUTierQuota(Base):
         index=True,
     )
     inference_name = Column(String(64), nullable=False)
-    monthly_quota = Column(BigInteger, nullable=False)
-    pending_monthly_quota = Column(BigInteger, nullable=True)
+    monthly_quota = Column(Numeric(15, 4), nullable=False)
+    pending_monthly_quota = Column(Numeric(15, 4), nullable=True)
     created_by = Column(String(255), nullable=True)
     updated_by = Column(String(255), nullable=True)
     created_at = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
